+++ src/mcp/index.ts
@@ -188,9 +188,6 @@ export class FirebaseMcpServer {
       return this.emulatorHubClient;
     }
     const projectId = await this.getProjectId();
-    if (!projectId) {
-      return;
-    }
     this.emulatorHubClient = new EmulatorHubClient(projectId);
     return this.emulatorHubClient;
   }