+++ CHANGELOG.md
@@ -1,2 +1,4 @@
 - Support auth token JSON in `dataconnect_execute` MPC tool. (#9046)
 - feat: [mcp] Add 'select' parameter to the `firestore_get_documents` tool. (#9043)
+- `firebase emulator:start` use a default project if no project can be found. (#9072)
+